requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
decorator==5.1.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.0
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pbr==5.11.1
peewee==3.16.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2022.12.19
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
stevedore==5.0.0
traitlets==5.9.0
typing_extensions==4.5.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
